Factoring make_realization_vector, which returns a new realization data_vector and make_realization which generate a new sacc with or without noise.
